Resuming active development. Now uses ternary operator wherever possible. Instance and Backup classes in their own file.
