+++ QBBot.py
@@ -1,6 +1,6 @@
 # Lev's Quizbowl Bot
 # Author: Lev Bernstein
-# Version: 1.8.12
+# Version: 1.8.13
 # This bot is designed to be a user-friendly Quizbowl Discord bot with a minimum of setup.
 # All commands are documented; if you need any help understanding them, try the command !tutorial.
 # This bot is free software, licensed under the GNU GPL version 3. If you want to modify the bot in any way,
@@ -29,11 +29,10 @@ from discord.utils import get
 
 # Custom modules:
 from Summon import *
-
+from Instance import *
 
 
 # Setup
-token = ""
 try:
     with open("token.txt", "r") as f: # in token.txt, paste in your own Discord API token
         token = f.readline()
@@ -52,9 +51,7 @@ def isInt(st):
         return False
     if st.startswith('<:ten:') or st.startswith('<:neg:') or st.startswith('<:power:'): # this conditional handles awarding points with emojis
         return True
-    if st[0] == '-' or st[0] == '+':
-        return st[1:].isdigit()
-    return st.isdigit()
+    return st[1:].isdigit() if (st[0] == '-' or st[0] == '+') else st.isdigit()
 
 def isBuzz(st):
     """Checks if an entered string is a valid buzz. If you want to allow additional means of buzzing, add them to validBuzzes."""
@@ -65,10 +62,8 @@ def isBuzz(st):
         st.startswith('!bz'),
         st.startswith('!buz'),
         st.startswith('<:bee:')
-        )
-    if any(validBuzzes):
-        return True
-    return False
+        )  
+    return any(validBuzzes)
 
 def writeOut(generate, name, content, game, report, spoke):
     """Saves output of valid commands in the log file.
@@ -85,462 +80,8 @@ def writeOut(generate, name, content, game, report, spoke):
                 f.write(newline)
     return
 
-#############################################################
-# Backup(self, prev)
-# This class is a backup of previous scores, for the purpose of being able to undo an incorrect points assignment.
-# It is essentially a singly linked list with some additional data fields.
-# Parameters: prev is the previous backup of the scores.
-# Local variables that require explaining:
-#   line: a backup of the previous bottom line of the scoresheet.
-# Will be used starting in Version 1.9.X.
-class Backup:
-    def __init__(self, prev):
-        # TODO store TUnum here as well, to better handle negs
-        self.scores = {}
-        self.redBonus = 0
-        self.blueBonus = 0
-        self.greenBonus = 0
-        self.orangeBonus = 0
-        self.yellowBonus = 0
-        self.purpleBonus = 0
-        self.TUnum = 0
-        self.lastNeg = False
-        self.prev = prev
-        self.line = ""
 
-#############################################################
-# Instance(self, channel)
-# This class is an instance of an active game of Quizbowl.
-# Every channel in which a game is run gets its own instance.
-# You cannot have more than one Instance per channel.
-# Parameters: channel is the id of the Discord channel in which the instance is run.
-# Local variables:
-#   Tunum: Represents the number of completed TUs read. Advances after points are awarded for a TU or a TU goes dead (!dead).
-#   scores: Dictionary mapping players to the number of points they have.
-#   buzzes: Queue containing players' buzzes. After a player negs, they are popped from the queue and it moves on to the next buzz.
-#   buzzed: Queue containing players who have already buzzed. If a player is in this queue, they cannot buzz again. You can clear this with !clear (or !dead, which also advances Tunum).
-#   active: Boolean that tracks whether there are currently valid buzzes on a TU.
-#   reader: Discord Member who is reading/moderating a given Instance.
-#   bonusEnabled: Boolean that controls whether the bot will prompt for bonus points. Can be toggled with !bonusmode.
-#   bonusMode: Boolean that tracks if reader is currently reading a bonus. Automatically set upon points being awarded for a TU or bonus.
-#   lastBonusMem: Discord Member who last gained points on a TU. If they are on a team, whatever they get on the bonus is given to their team. Otherwise, it's added to their score.
-#   red/blue/.../purpleTeam: Arrays of Discord Members from each team.
-#   red/blue/.../purpleNeg: Booleans tracking if a team is locked out from buzzing due to a member of their team negging.
-#   red/blue/.../purpleBonus: The number of points each team has earned on bonuses so far.
-#   logFile: A log created to track all commands from this particular game. Untracked by .gitignore.
-#   lastTossupPoints: Used for giving bonus points to the individual when playing without teams.
-#   lastNeg: Boolean that tracks whether the last buzz was a 0/-5. Used in undo.
-#   oldScores: A Backup() of a given Instance's scores.
-class Instance: # instance of an active game. Each channel a game is run in gets its own instance. You cannot have more than one game per channel.
-    def __init__(self, channel):
-        self.channel = channel
-        self.TUnum = 0
-        self.scores = {}
-        self.buzzes = deque()
-        self.buzzed = deque()
-        self.active = False
-        self.reader = None
-        self.bonusEnabled = True
-        self.bonusMode = False
-        self.lastBonusMem = None
-        self.redTeam = []
-        self.blueTeam = []
-        self.greenTeam = []
-        self.orangeTeam = []
-        self.yellowTeam = []
-        self.purpleTeam = []
-        self.redNeg = False
-        self.blueNeg = False
-        self.greenNeg = False
-        self.orangeNeg = False
-        self.yellowNeg = False
-        self.purpleNeg = False
-        self.redBonus = 0
-        self.blueBonus = 0
-        self.greenBonus = 0
-        self.orangeBonus = 0
-        self.yellowBonus = 0
-        self.purpleBonus = 0
-        self.logFile = ("gamelogs/" + str(self.getChannel())[-5:] + "-" + datetime.now().strftime("%Y-%m-%d-%H-%M-%S") + ".log")
-        self.csvScore = ("gamelogs/" + str(self.getChannel())[-5:] + "-" + datetime.now().strftime("%Y-%m-%d-%H-%M-%S") + ".csv")
-        # log and scoresheet filename format: channelID-YYYY-mm-DD-HH-MM-SS
-        self.lastTossupPoints = 0
-        # Used starting in Version 1.9.X:
-        self.lastNeg = False
-        self.oldScores = Backup(None)
-
-    def getChannel(self):
-        """Return the channel of a given Instance. 
-        TODO For the sake of proper encapsulation, to make my CS101 professor proud, I should switch to method calls like this for every local variable I access in on_message().
-        """
-        return self.channel
-    
-    def buzz(self, mem):
-        """Adds a buzzing player to the buzzes and buzzed arrays, if they are allowed to buzz."""
-        if self.hasBuzzed(mem):
-            print("Extra")
-        else:
-            self.active = True
-            self.buzzes.append(mem)
-            self.buzzed.append(mem)
-            print("Appended")
-    
-    def hasBuzzed(self, mem):
-        """Returns whether a player has already buzzed. Returns True if they have already buzzed, and cannot do so again; False otherwise."""
-        if mem in self.buzzed:
-            return True
-        return False
-    
-    def clear(self):
-        """Clears the buzzes and buzzed arrays, allowing all players to buzz in once more."""
-        print("Clearing...")
-        self.buzzes = deque()
-        self.buzzed = deque()
-        self.unlock()
-        self.active = False
-    
-    def dead(self):
-        """Marks a Tossup as dead, clears the buzzes and buzzed arrays, and advances TUnum."""
-        temp = copy.copy(self.oldScores)
-        self.oldScores.prev = temp
-        self.oldScores.scores = copy.copy(self.scores)
-        self.oldScores.redBonus = self.redBonus
-        self.oldScores.blueBonus = self.blueBonus
-        self.oldScores.greenBonus = self.greenBonus
-        self.oldScores.orangeBonus = self.orangeBonus
-        self.oldScores.yellowBonus = self.yellowBonus
-        self.oldScores.purpleBonus = self.purpleBonus
-        self.oldScores.TUnum = self.TUnum
-        self.oldScores.lastNeg = self.lastNeg
-        self.clear()
-        self.lastNeg = False
-        self.TUnum +=1
-    
-    def unlock(self):
-        """Allows all teams to buzz in again."""
-        print("Unlocking...")
-        self.redNeg = False
-        self.blueNeg = False
-        self.greenNeg = False
-        self.orangeNeg = False
-        self.yellowNeg = False
-        self.purpleNeg = False
-    
-    def canBuzz(self, mem):
-        """Checks if a player's team is allowed to buzz.
-        This exists alongside hasBuzzed() to reflect the two different reasons for not being allowed to buzz:
-        Having personally negged, or a teammate having negged. canBuzz() checks if a teammate has negged.
-        """
-        conditions = (
-            (mem in self.redTeam and self.redNeg==True),
-            (mem in self.blueTeam and self.blueNeg==True),
-            (mem in self.greenTeam and self.greenNeg==True),
-            (mem in self.orangeTeam and self.orangeNeg==True),
-            (mem in self.yellowTeam and self.yellowNeg==True),
-            (mem in self.purpleTeam and self.purpleNeg==True)
-        )
-        if any(conditions):
-            return False
-        return True
-    
-    def inTeam(self, text, mem):
-        """When bonus mode is enabled, this method reports the team of the buzzer."""
-        if mem in self.redTeam:
-            role = get(text.guild.roles, name = 'Team red')
-            return role
-        if mem in self.blueTeam:
-            role = get(text.guild.roles, name = 'Team blue')
-            return role
-        if mem in self.greenTeam:
-            role = get(text.guild.roles, name = 'Team green')
-            return role
-        if mem in self.orangeTeam:
-            role = get(text.guild.roles, name = 'Team orange')
-            return role
-        if mem in self.yellowTeam:
-            role = get(text.guild.roles, name = 'Team yellow')
-            return role
-        if mem in self.purpleTeam:
-            role = get(text.guild.roles, name = 'Team purple')
-            return role
-        return None
-    
-    def teamScore(self, team, teamBonus):
-        """Returns a team's total score, including bonus points."""
-        total = teamBonus
-        for mem in team:
-            if mem in self.scores:
-                total += self.scores[mem]
-        return total
-    
-    def teamExist(self, team):
-        """Checks if a given team has members"""
-        if len(team) > 0:
-            return True
-        return False
-    
-    """
-    def undo(self):
-        # Reverts scores back to one tossup ago. Currently has major problems; see Issue #7 on the GitHub Issues page.
-        self.scores = copy.copy(self.oldScores.scores)
-        self.redBonus = self.oldScores.redBonus
-        self.blueBonus = self.oldScores.blueBonus
-        self.greenBonus = self.oldScores.greenBonus
-        self.orangeBonus = self.oldScores.orangeBonus
-        self.yellowBonus = self.oldScores.yellowBonus
-        self.purpleBonus = self.oldScores.purpleBonus
-        self.oldScores = self.oldScores.prev
-        self.TUnum = self.oldScores.TUnum
-        if self.lastNeg:
-            self.TUnum += 1
-            self.oldScores.TUnum += 1
-        self.lastNeg = self.oldScores.lastNeg
-        self.clear()
-    """
-    
-    def gain(self, points):
-        """Awards points to the player at the front of the buzzes queue. 
-        If the points awarded is a positive number, they have gotten a TU correct, so it moves on to a bonus if those are enabled. Returns true.
-        If the points awarded is a negative number or 0, they have gotten a TU incorrect, so it adds that to the individual's score and does not advance TUnum. Returns false.
-        Rarely, a pickling error arises with the backup system; might have fixed by switching to copy from deepcopy.
-        """
-        awarded = False
-        if self.active == True:
-            self.lastTossupPoints = points
-            mem = self.buzzes.popleft()
-            temp = copy.copy(self.oldScores)
-            self.oldScores.prev = temp
-            self.oldScores.scores = copy.copy(self.scores)
-            self.oldScores.redBonus = self.redBonus
-            self.oldScores.blueBonus = self.blueBonus
-            self.oldScores.greenBonus = self.greenBonus
-            self.oldScores.orangeBonus = self.orangeBonus
-            self.oldScores.yellowBonus = self.yellowBonus
-            self.oldScores.purpleBonus = self.purpleBonus
-            self.oldScores.TUnum = self.TUnum
-            self.oldScores.lastNeg = self.lastNeg
-            if points > 0: # Someone got a TU correct.
-                self.lastNeg = False
-                if mem in self.scores:
-                    self.scores[mem] = self.scores[mem] + points
-                    self.active = False
-                    self.clear()
-                    awarded = True
-                else:
-                    #with open(self.csvScore, "r") as f:
-                        #body = f.readlines()
-                    #print("Body 0 = " + body[0])
-                    #lane = body[0]
-                    #test = lane.split(",")
-                    #if mem.name not in test:
-                        #newLane = "TU#,Red Bonus,Blue Bonus,Green Bonus,Orange Bonus,Yellow Bonus,Purple Bonus,"
-                        #for x,y in self.scores.items():
-                            #newLane += x.name + ","
-                        #newLane += mem.name + ",\r\n"
-                        #print("New Body 0 = " + newLane)
-                        #body = ''.join([i for i in body]) \
-                            #.replace(lane, newLane)
-                        #print(body)
-                        #with open(self.csvScore, "w") as f:
-                            #f.writelines(body)
-                    self.scores[mem] = points
-                    self.active = False
-                    self.clear()
-                    awarded = True
-                if self.bonusEnabled:
-                    self.lastBonusMem = mem
-                    self.bonusMode = True # Move on to awarding bonus points
-            else: #Someone got a 0 or a -5.
-                self.lastNeg = True
-                if mem in self.scores:
-                    self.scores[mem] = self.scores[mem] + points
-                else:
-                    self.scores[mem] = points
-                    #with open(self.csvScore, "r") as f:
-                        #body = f.readlines()
-                    #print("Body 0 = " + body[0])
-                    #lane = body[0]
-                    #test = lane.split(",")
-                    #if mem.name not in test:
-                        #newLane = "TU#,Red Bonus,Blue Bonus,Green Bonus,Orange Bonus,Yellow Bonus,Purple Bonus,"
-                        #for x,y in self.scores.items():
-                            #newLane += x.name + ","
-                        #newLane += mem.name + ",\r\n"
-                        #print("New Body 0 = " + newLane)
-                        #body = ''.join([i for i in body]) \
-                            #.replace(lane, newLane)
-                        #print(body)
-                        #with open(self.csvScore, "w") as f:
-                            #f.writelines(body)
-                if len(self.buzzes) == 0:
-                    self.active = False
-                if mem in self.redTeam:
-                    self.redNeg = True
-                    print("red locked out")
-                if mem in self.blueTeam:
-                    self.blueNeg = True
-                    print("blue locked out")
-                if mem in self.greenTeam:
-                    self.greenNeg = True
-                    print("green locked out")
-                if mem in self.orangeTeam:
-                    self.orangeNeg = True
-                    print("orange locked out")
-                if mem in self.yellowTeam:
-                    self.yellowNeg = True
-                    print("yellow locked out")
-                if mem in self.purpleTeam:
-                    self.purpleNeg = True
-                    print("purple locked out")
-            #with open(self.csvScore) as f:
-                #body = f.readlines()
-                #subMems = body[0].split(',')
-                #found = False
-                #for i in range(len(subMems)):
-                    #if mem.name == subMems[i]:
-                        #spot = i
-                        #found = True
-                        #break
-            #if found:
-                #if self.oldScores.lastNeg: # if there has already been a neg on the current TU:
-                    #with open(self.csvScore, "r") as f :# split the last line; replace spot with points
-                        #reader = csv.reader(f, delimiter=',')
-                        #count = 0
-                        #for row in reader:
-                            #count +=1
-                    #print(count)
-                    #try:
-                        #oldLane = body[count]
-                        #newLane = body[count].split(',')
-                        #print(newLane)
-                        #print("New line length: " + str(len(newLane)))
-                        #newLane[spot] = str(points)
-                        #total = ""
-                        #for item in newLane:
-                            #total += item + ","
-                        #body = ''.join([i for i in body]) \
-                            #.replace(oldLane, total)
-                        #with open(self.csvScore, "w") as f:
-                            #f.writelines(body)
-                    #except IndexError as err: # TODO fix indexerror that pops up from time to time
-                        #print("IndexError!")
-                        #print(err)
-                #else:
-                    #newLine = [str(self.TUnum + 1)]
-                    #for i in range(6):
-                        #newLine.append('0')
-                    #for i in range (spot-7):
-                        #newLine.append('')
-                    #newLine.append(str(points))
-                    #print(newLine)
-                    #with open(self.csvScore, "a+", newline='') as f:
-                        #writer = csv.writer(f)
-                        #writer.writerow(newLine)
-        if awarded:
-            self.TUnum +=1 # If a positive # of points has been assigned, that means someone got the TU correct. Advance the TU count.
-        return awarded
-
-    def bonusGain(self, points):
-        """Awards bonus points, either to the team or to the individual if playing without teams."""
-        if not self.bonusEnabled:
-            return
-        if not self.bonusMode:
-            return
-        temp = copy.copy(self.oldScores)
-        conditions = (
-            self.lastBonusMem in self.redTeam,
-            self.lastBonusMem in self.blueTeam,
-            self.lastBonusMem in self.greenTeam,
-            self.lastBonusMem in self.orangeTeam,
-            self.lastBonusMem in self.yellowTeam,
-            self.lastBonusMem in self.purpleTeam
-            )
-        changed = -1
-        if any(conditions):
-            if self.lastBonusMem in self.redTeam:
-                self.redBonus += points
-                changed = 0
-            if self.lastBonusMem in self.blueTeam:
-                self.blueBonus += points
-                changed = 1
-            if self.lastBonusMem in self.greenTeam:
-                self.greenBonus += points
-                changed = 2
-            if self.lastBonusMem in self.orangeTeam:
-                self.orangeBonus += points
-                changed = 3
-            if self.lastBonusMem in self.yellowTeam:
-                self.yellowBonus += points
-                changed = 4
-            if self.lastBonusMem in self.purpleTeam:
-                self.purpleBonus += points
-                changed = 5
-        else:
-            self.scores[self.lastBonusMem] += points
-        #with open(self.csvScore, "r+") as f:
-            #body = f.readlines()
-            #lastLine = body.pop().split(',')
-        #with open(self.csvScore, "w") as f:
-            #f.writelines(body)
-        #print(lastLine)
-        #lastLine[1] = "0"
-        #lastLine[2] = "0"
-        #lastLine[3] = "0"
-        #lastLine[4] = "0"
-        #lastLine[5] = "0"
-        #lastLine[6] = "0"
-        #searching = True
-        #if changed != -1:
-            #if searching and changed == 0:
-                #lastLine[1] = str(points)
-                #searching = False
-            #if searching and changed == 1:
-                #lastLine[2] = str(points)
-                #searching = False
-            #if searching and changed == 2:
-                #lastLine[3] = str(points)
-                #searching = False
-            #if searching and changed == 3:
-                #lastLine[4] = str(points)
-                #searching = False
-            #if searching and changed == 4:
-                #lastLine[5] = str(points)
-                #searching = False
-            #if searching and changed == 5:
-                #lastLine[6] = str(points)
-        #else:
-            #print("selfAdded")
-            #found = False
-            #with open(self.csvScore) as f:
-                #body = f.readlines()
-                #subMems = body[0].split(',')
-                #found = False
-                #for i in range(len(subMems)):
-                    #if self.lastBonusMem.name == subMems[i]:
-                        #spot = i
-                        #found = True
-                        #break
-            #if found:
-                #print("found")
-                #print(self.lastTossupPoints)
-                #print(points)
-                #lastLine[spot] = str(self.lastTossupPoints + points)
-                #print(lastLine[spot])
-        #with open(self.csvScore, "a+", newline='') as f:
-            #writer = csv.writer(f)
-            #print(lastLine)
-            #writer.writerow(lastLine)
-        self.bonusMode = False
-        
-    def bonusStop(self):
-        """Kills an active bonus."""
-        if self.bonusEnabled:
-            self.lastBonusMem = None
-            self.bonusMode = False
-        else:
-            print("Could not stop a bonus!")
-    
-games = [] # Array holding all active games across all channels and servers. If this bot sees sufficiently large usage, I will switch to something more sorted to store the games, like a binary tree or a 2-3 tree. Games would be sorted based on channel ID.
+games = [] # List holding all active games across all channels and servers. If this bot sees sufficiently large usage, I will switch to something more sorted to store the games, like a binary tree or a 2-3 tree. Games would be sorted based on channel ID.
 
 
 @client.event
@@ -616,7 +157,7 @@ async def on_message(text):
                     with open("templates/emoji/power.png", "rb") as powerIMG:
                         img = powerIMG.read()
                         power = await text.guild.create_custom_emoji(name = 'power', image = img)
-                    print("Emojis live!")
+                    print("Emoji live!")
                     report += " Team roles now exist, as do the following emoji: " + str(buzz) + ", " + str(neg) + ", " + str(ten) + ", " + str(power) + "."
                     with open("templates/pfp.png", "rb") as pfp:
                         pic = pfp.read()
@@ -640,9 +181,7 @@ async def on_message(text):
             botSpoke = True
             report = "This command is only usable by server admins!"
             if text.author.guild_permissions.administrator: # this makes sure people can't just ping everyone in the server whenever they want. Only admins can do that.
-                report = "@everyone it's time for Quiz Bowl practice."
-                if verbosePings:
-                    report = summon() # For the full list of summon messages, check Summon.py.
+                report = summon() if verbosePings else "@everyone it's time for Quiz Bowl practice."# For the full list of summon messages, check Summon.py.
             await text.channel.send(report)
             if exist:
                 writeOut(generateLogs, text.author.name, text.content, heldGame, report, botSpoke)
@@ -774,9 +313,9 @@ async def on_message(text):
                 print(text.content + " is an int")
                 if text.content.startswith('<:neg:'): # This and the next two conditionals check to see if someone is using a valid emoji to assign points. While, to the user, an emoji looks like :emojiName:, to the bot it is also wrapped in <>.
                     text.content = "-5"
-                if text.content.startswith('<:ten:'):
+                elif text.content.startswith('<:ten:'):
                     text.content = "10"
-                if text.content.startswith('<:power:'):
+                elif text.content.startswith('<:power:'):
                     text.content = "15"
                 botSpoke = True
                 report = "Null"
@@ -817,7 +356,7 @@ async def on_message(text):
                                     else:
                                         heldGame.buzzes.popleft() # Pop until we find someone who can buzz, or until the array of buzzes is empty.
                                         report = "Cannot buzz."
-                                        # because I don't send a report here, I can't have the text.channel.send(report) at the end; doing so throws an exception because it would be sending an empty message
+                                        # because I don't want to send a report here, I can't have the text.channel.send(report) at the end
                         else: # bonusMode true
                             heldGame.bonusGain(int(text.content))
                             report = "Awarded bonus points. Moving on to TU #" + str(heldGame.TUnum + 1) + "."
@@ -1034,9 +573,6 @@ async def on_message(text):
             report = "Current TU: #" + str(heldGame.TUnum + 1) + "."
             await text.channel.send(report)
             writeOut(generateLogs, text.author.name, text.content, heldGame, report, botSpoke)
-            return
-            if exist:
-                writeOut(generateLogs, text.author.name, text.content, heldGame, report, botSpoke)
             return
         
         if text.content.startswith('!team'):
